+++ game/cell.py
@@ -1,0 +1,7 @@
+
+    ### State = True if theres no piece or False if its occupied
+    ### Piece refers to the object
+class Cell():
+    def __init__(self, state, piece):
+        self.__state__= state
+        self.__piece__ = piece