+++ CHANGELOG.md
@@ -4,3 +4,7 @@
 - Created pieces with color and position on the board
 - Created class Board, with print_board(basic)
 - Created Cell, with state and piece
+- Created class for each piece, all of them are sons of class Pieces
+
+# DELETED
+- Everything from branch main